+++ index.js
@@ -296,6 +296,7 @@ class MelCloudPlatform {
 						clientId: mqtt.clientId ? `melcloud_${mqtt.clientId}_${Math.random().toString(16).slice(3)}` : `melcloud_${Math.random().toString(16).slice(3)}`,
 						prefix: mqtt.prefix ? `melcloud/${mqtt.prefix}/${deviceTypeString}/${deviceName}` : `melcloud/${deviceTypeString}/${deviceName}`,
 						protocolVersion: mqtt.protocolVersion,
+						haDiscovery: mqtt.haDiscovery,
 						user: mqtt.auth?.user,
 						passwd: mqtt.auth?.passwd,
 						logWarn: logLevel.warn,